backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
psycopg2-binary
pyjwt==2.8.0
python-dotenv==1.0.0
pydantic>=2.5.0
python-multipart==0.0.9
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
email-validator>=2.1.0